Clear account cache if file exist.

Former-commit-id: a36019ea357088f0c80e0a07f7c0637fc04c6d0d [formerly 4c0ed0fef903af870f69a8fc819eacd386a11167]
Former-commit-id: 85acd7072c727d19bcc097ea1ae02df69a446be4 [formerly 36545e0c85bec599ad5ef3512df8808681203004]
Former-commit-id: 3fe030077faa1a3b95ffbe2bc7083af94ccc3560

+++ Src/AppLib/VpnHood.AppLib.App/Services/Accounts/AppAccountService.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json;
+﻿using System.Net;
+using System.Text.Json;
 using VpnHood.AppLib.Abstractions;
 using VpnHood.Core.Common.Logging;
 using VpnHood.Core.Common.Utils;
@@ -67,7 +68,8 @@ public class AppAccountService
 
     private void ClearCache()
     {
-        File.Delete(AppAccountFilePath);
+        if (File.Exists(AppAccountFilePath))
+            File.Delete(AppAccountFilePath);
         _appAccount = null;
     }
